% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scLS.plot.R
\name{scLS.plot}
\alias{scLS.plot}
\title{Plot Lomb–Scargle periodogram for a single feature}
\usage{
scLS.plot(
  object,
  time.col,
  feature,
  assay = "RNA",
  slot = "data",
  center = FALSE,
  window.func = NULL,
  f.min = 0,
  f.max = 2,
  n.bins = 500
)
}
\arguments{
\item{object}{Seurat object}

\item{time.col}{Numeric metadata column name containing sampling times}

\item{feature}{Single gene/feature name to plot}

\item{assay}{Assay name (default "RNA")}

\item{slot}{Slot name (default "data")}

\item{center}{Logical; if TRUE, subtract the mean after windowing}

\item{window.func}{String name of window function (e.g. "hanning") or NULL}

\item{f.min}{Numeric; minimum frequency}

\item{f.max}{Numeric; maximum frequency}

\item{n.bins}{Integer; number of frequency bins}
}
\value{
A ggplot object of frequency vs. power
}
\description{
Plot Lomb–Scargle periodogram for a single feature
}
